remove outdated consts, put DEFAULT_BOUNDS in shapes file
